% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyDigitise_functions.R
\name{check_calibrate}
\alias{check_calibrate}
\title{check_calibration}
\usage{
check_calibrate(x)
}
\arguments{
\item{x}{list data created in sDigitise about the graph}
}
\description{
Checks whether a graph has been calibrated
}
