% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyDigitise_functions.R
\name{check_extract}
\alias{check_extract}
\title{check_extract}
\usage{
check_extract(x)
}
\arguments{
\item{x}{list data created in sDigitise about the graph}
}
\description{
Checks whether a graph has been extracted
}
