% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyDigitise_functions.R
\name{check_orientation}
\alias{check_orientation}
\title{check_orientation}
\usage{
check_orientation(x)
}
\arguments{
\item{x}{list data created in sDigitise about the graph}
}
\description{
Checks whether a graph has been orientated
}
