% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyDigitise_functions.R
\name{check_plottype}
\alias{check_plottype}
\title{check_plottype}
\usage{
check_plottype(x)
}
\arguments{
\item{x}{list data created in sDigitise about the graph}
}
\description{
Checks whether the plottype has been selected
}
