% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaDigitise_functions.R
\name{internal_redraw}
\alias{internal_redraw}
\title{internal_redraw}
\usage{
internal_redraw(
  image_file,
  flip = FALSE,
  rotate = 0,
  plot_type = NULL,
  variable = NULL,
  cex = NULL,
  calpoints = NULL,
  point_vals = NULL,
  raw_data = NULL,
  rotation = TRUE,
  calibration = TRUE,
  points = TRUE,
  rotate_mode = FALSE,
  pos = NULL,
  shiny = FALSE,
  zoom_coords = NULL,
  ...
)
}
\arguments{
\item{image_file}{Image filename}

\item{flip}{whether to flip figure}

\item{rotate}{how much to rotate figure}

\item{plot_type}{plot_type}

\item{variable}{variable}

\item{cex}{relative size of points and text}

\item{calpoints}{The calibration points}

\item{point_vals}{The point values}

\item{raw_data}{The raw data}

\item{rotation}{logical, should figure be rotated}

\item{calibration}{logical, should calibration be redrawn}

\item{points}{logical, should points be redrawn}

\item{rotate_mode}{logical, has the plot been rotated}

\item{pos}{legend position}

\item{shiny}{logical if using Shiny or not}

\item{zoom_coords}{the cordinates for zooming}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Redraws figure and extraction data (a metaDigitise function)
}
