% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaDigitise_functions.R
\name{process_data}
\alias{process_data}
\title{process data}
\usage{
process_data(object)
}
\arguments{
\item{object}{object data with plot type and variables}
}
\description{
Processing data (a metaDigitise function)
}
