% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaDigitise_functions.R
\name{redraw_calibration}
\alias{redraw_calibration}
\title{metaDigitise_functions}
\usage{
redraw_calibration(
  plot_type,
  variable,
  calpoints,
  point_vals,
  image_details,
  cex
)
}
\description{
Various metaDigitise functions
}
