% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaDigitise_functions.R
\name{convert_histogram_data}
\alias{convert_histogram_data}
\title{convert histogram data}
\usage{
convert_histogram_data(cal_data)
}
\arguments{
\item{cal_data}{Calibration data}
}
\value{
converted histogram data
}
\description{
Conversion of histogram data by metaDigitise
}
