% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaDigitise_functions.R
\name{internal_redraw}
\alias{internal_redraw}
\title{metaDigitise_functions}
\usage{
internal_redraw(
  image_file,
  flip = FALSE,
  rotate = 0,
  plot_type = NULL,
  variable = NULL,
  cex = NULL,
  calpoints = NULL,
  point_vals = NULL,
  raw_data = NULL,
  rotation = TRUE,
  calibration = TRUE,
  points = TRUE,
  rotate_mode = FALSE,
  pos = NULL,
  shiny = FALSE,
  zoom_coords = NULL,
  ...
)
}
\description{
Various metaDigitise functions
}
